solving issues for pull request
